refactor: moved svgs into a separate file
