+++ frontend/src/components/BookmarkList/BookmarkList.js
@@ -1,5 +1,6 @@
 import { useContext, useEffect } from "react";
 import { Context } from "../../Context";
+import { removeBookmarkIcon } from "../../svgs";
 
 function BookmarkList() {
   const { bookmarks, setBookmarks, removeBookmark, setPageTitle } =
@@ -36,18 +37,7 @@ function BookmarkList() {
               onClick={() => removeBookmark(bookmark.url)}
               title="remove bookmark"
             >
-              <svg
-                aria-hidden="true"
-                role="img"
-                width="32"
-                height="32"
-                preserveAspectRatio="xMidYMid meet"
-                viewBox="0 0 16 16"
-              >
-                <g fill="currentColor">
-                  <path d="M2 2v13.5a.5.5 0 0 0 .74.439L8 13.069l5.26 2.87A.5.5 0 0 0 14 15.5V2a2 2 0 0 0-2-2H4a2 2 0 0 0-2 2z"></path>
-                </g>
-              </svg>
+              {removeBookmarkIcon}
             </button>
           </div>
         );
